// drone-debuggers-main/frontend/src/case-form-page/models/enums/disruption-option.enum.ts
export enum DisruptionOption {
  MoreThan14Days = '>14 days',
  LessThan14Days = '<14 days',
  OnFlightDay = 'on flight day',
  MoreThan3Hours = '>3 hours',
  LessThan3Hours = '<3 hours',
  NeverArrived = 'Never Arrived',
  FlightOverbooked = 'Flight Overbooked',
  AggressiveBehaviour = 'Aggressive Behaviour',
  Intoxication = 'Intoxication',
  UnspecifiedReason = 'Unspecified Reason'
}